refactor: move expressive code style overrides to be inline
